+++ src/@types/declarations.d.ts
@@ -1,1 +1,2 @@
 declare module "*.scss";
+declare module "*.png";
