+++ src/pages/MyPage/Avatar/style.module.scss
@@ -1,5 +1,4 @@
 @import "src/assets/modules";
-
 .root {
   .uploadIcon {
     background-size: contain;
